Safari-609.1.11.1/LayoutTests/fast/canvas/canvas-strokeRect-lineJoin-expected.txt

Ensure correct behavior of canvas with stroke rect with line join

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS data[0] is 0
PASS data[0] is 0
PASS data[0] is 0
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 255
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 255
PASS data[0] is not 255
PASS data[0] is not 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 0
PASS data[0] is 0
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 255
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 255
PASS data[0] is not 255
PASS data[0] is not 255
PASS data[0] is 0
PASS data[0] is 255
PASS data[0] is 0
PASS data[0] is 255
PASS successfullyParsed is true

TEST COMPLETE
